+++ client/my-sites/plans-features-main/components/plan-upsell-modal/free-plan-paid-domain-dialog.tsx
@@ -19,7 +19,6 @@ import { DomainPlanDialogProps, MODAL_VIEW_EVENT_NAME } from '.';
 export function FreePlanPaidDomainDialog( {
 	paidDomainName,
 	generatedWPComSubdomain,
-	suggestedPlanSlug,
 	onFreePlanSelected,
 	onPlanSelected,
 }: DomainPlanDialogProps ) {
